Conflicts:
	src/Blocks/WorldInterface.h
	src/ClientHandle.cpp
	src/ClientHandle.h
	src/Entities/Player.cpp
	src/Entities/Player.h
	src/Generating/FinishGen.cpp
	src/Protocol/Protocol.h
	src/Protocol/Protocol125.cpp
	src/Protocol/Protocol125.h
	src/Protocol/Protocol16x.cpp
	src/Protocol/Protocol16x.h
	src/Protocol/Protocol17x.cpp
	src/Protocol/Protocol17x.h
	src/Protocol/ProtocolRecognizer.cpp
	src/Protocol/ProtocolRecognizer.h
	src/Root.h
	src/World.cpp

+++ src/Mobs/Zombie.h
@@ -17,7 +17,7 @@ public:
 	CLASS_PROTODEF(cZombie);
 	
 	virtual void GetDrops(cItems & a_Drops, cEntity * a_Killer = NULL) override;
-	virtual void MoveToPosition(const Vector3f & a_Position) override;
+	virtual void MoveToPosition(const Vector3d & a_Position) override;
 
 	bool IsVillagerZombie(void) const {return m_IsVillagerZombie; }
 	bool IsConverting    (void) const {return m_IsConverting; }
